requirements.txt
fastapi
uvicorn
python-dotenv
unstructured
unstructured_inference
pypdf
pdfminer.six
pillow
pdf2image
faiss-cpu
langchain
langchain-core
langchain-openai
langchain-community
openai
tiktoken
duckduckgo-search
pi_heif
unstructured_pytesseract
streamlit
pytesseract